Decision tree code + examples wrapper class
Decision tree code and wrapper class for feature vectors. Decision tree takes an "Examples" object (list of "Example" objects, which are feature vectors), and constructs a tree recursively using the standard ID3 algorithm. Follows Russel and Norvig pseudocode outline more or less line for line. Code will need to be changed to conform with the requirements of the coursework, but provides a good template.